Front fixes. Uninstall fixes

+++ uninstall.php
@@ -16,27 +16,29 @@ function uninstall()
 	$wpdb->query("DELETE FROM {$wpdb->options} WHERE option_name LIKE 'titan_%';");
 }
 
-if( is_multisite() ) {
-	global $wpdb, $wp_version;
+if(get_option( 'titan_complete_uninstall', false)) {
+	if ( is_multisite() ) {
+		global $wpdb, $wp_version;
 
-	$wpdb->query("DELETE FROM {$wpdb->sitemeta} WHERE meta_key LIKE 'titan_%';");
+		$wpdb->query("DELETE FROM {$wpdb->sitemeta} WHERE meta_key LIKE 'titan_%';");
 
-	$blogs = $wpdb->get_col("SELECT blog_id FROM $wpdb->blogs");
+		$blogs = $wpdb->get_col( "SELECT blog_id FROM $wpdb->blogs" );
 
-	if( !empty($blogs) ) {
-		foreach($blogs as $id) {
+		if ( ! empty( $blogs ) ) {
+			foreach ( $blogs as $id ) {
 
-			switch_to_blog($id);
+				switch_to_blog( $id );
 
-			uninstall();
+				uninstall();
 
-			restore_current_blog();
+				restore_current_blog();
+			}
 		}
+	} else {
+		uninstall();
 	}
-} else {
-	uninstall();
 }
 
-require_once dirname(__FILE__) . '/includes/firewall/class-database-schema.php';
-$db_schema = new \WBCR\Titan\Database\Schema();
-$db_schema->dropAll();
+//require_once dirname(__FILE__) . '/includes/firewall/class-database-schema.php';
+//$db_schema = new \WBCR\Titan\Database\Schema();
+//$db_schema->dropAll();
